+++ framework/meta/src/generate_snippets/snippet_crate_gen.rs
@@ -3,6 +3,8 @@ use std::{
     io::Write,
 };
 
+static SNIPPETS_SOURCE_FILE_NAME: &str = "interact.rs";
+
 pub(crate) fn create_snippets_folder(snippets_folder_path: &str) {
     // returns error if folder already exists, so we ignore the result
     let _ = fs::create_dir(snippets_folder_path);
@@ -61,7 +63,7 @@ publish = false
 
 [[bin]]
 name = "rust-interact"
-path = "src/lib.rs"
+path = "src/{SNIPPETS_SOURCE_FILE_NAME}"
 
 [dependencies.{contract_crate_name}]
 path = ".."
@@ -93,7 +95,7 @@ pub(crate) fn create_src_folder(snippets_folder_path: &str) {
 
 #[must_use]
 pub(crate) fn create_and_get_lib_file(snippets_folder_path: &str, overwrite: bool) -> File {
-    let lib_path = format!("{snippets_folder_path}/src/lib.rs");
+    let lib_path = format!("{snippets_folder_path}/src/{SNIPPETS_SOURCE_FILE_NAME}");
     if overwrite {
         File::create(&lib_path).unwrap()
     } else {